+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 + \[BREAKING] Changed the behavior of `makelist()`: now it's also a decorator, converting its return type to a list (revertable with `wrap=False`)
 + New module `lex` with functions `symbol_table()` and `lex()` — make tokenization more affordable
 + Add `dorks` module and `flask.harden()`
-+ Added `addattr()`
++ Add `sqlalchemy.bool_column()`: make making flags painless
++ Added `addattr()`, `PrefixIdentifier()`
 
 ## 0.3.6